NewProject/requirements.txt
# AI 배달 도우미 프로토타입 필요 라이브러리

# 웹 프레임워크
fastapi==0.104.1
uvicorn==0.24.0

# Twilio (전화 수신/발신)
twilio==8.10.3

# OpenAI (GPT-4 대화)
openai==1.3.7

# 음성 처리
SpeechRecognition==3.10.0
gtts==2.4.0
pydub==0.25.1

# 오디오 처리
pygame==2.5.2
pyaudio==0.2.11

# 데이터베이스
sqlite3  # Python 기본 포함
sqlalchemy==2.0.23

# HTTP 요청
requests==2.31.0
httpx==0.25.2

# 환경 변수 관리
python-dotenv==1.0.0

# 데이터 처리
pandas==2.1.3
pyjwt==2.8.0

# 로깅
loguru==0.7.2

# 날짜/시간 처리
python-dateutil==2.8.2

# 텍스트 처리
regex==2023.10.3